Adding robust logreg handling for fNML and qNML (rethink file format). Adding initial support for qNML (not ready)

+++ src/cfg.h
@@ -8,6 +8,6 @@
 
 typedef unsigned int varset_t;
 
-typedef float score_t;
+typedef double score_t;
 
 #endif
